+++ NEWS.md
@@ -1,3 +1,8 @@
+# Version (development version)
+
+ * ...
+
+
 # Version 1.2.1 (2026-02-19)
 
 ## Bug Fixes