+++ NEWS.md
@@ -1,4 +1,4 @@
-# Version (development version)
+# Version 1.2.2 (2026-05-21)
 
 ## Bug Fixes